Roles permissions

+++ app/controllers/members_controller.rb
@@ -3,9 +3,9 @@ class MembersController < ApplicationController
 
   before_action :require_login
 
-  # authorize_actions_for Member
-
   def index
+    authorize Membership
+
     render locals: {
       members: current_account.memberships.ordered,
       invites: current_account.invites.ordered,
@@ -14,26 +14,30 @@ class MembersController < ApplicationController
   end
 
   def edit
-    # authorize_action_for member
+    authorize member
 
     render locals: { member: member }
   end
 
   def update
-    # authorize_action_for member, update_params
+    authorize member
 
-    member.update! update_params
+    member.update! permitted_params
     success_redirect
   rescue ActiveRecord::RecordInvalid => _err
     render 'edit'
   end
 
   def destroy
+    authorize member
+
     member.destroy!
     redirect_to :back
   end
 
   def send_email_confirmation
+    authorize member, :update?
+
     member.user.deliver_email_confirmation!
     redirect_to :back, flash: { success: I18n.t('flashes.user.confirmation_sent') }
   end
@@ -57,10 +61,6 @@ class MembersController < ApplicationController
   end
 
   def permitted_params
-    params.require(:invite).permit(:email)
-  end
-
-  def update_params
     params.require(:membership).permit(:role, :sms_notification, :email_notification)
   end
 end
